Moved action server to core
