+++ eager_bridge_gazebo/package.xml
@@ -42,9 +42,5 @@
   <exec_depend>gazebo_msgs</exec_depend>
   <exec_depend>roslaunch</exec_depend>
   <exec_depend>sensor_msgs</exec_depend>
-  <exec_depend>control_msgs</exec_depend>
-  <exec_depend>trajectory_msgs</exec_depend>
-  <exec_depend>actionlib</exec_depend> 
-  <exec_depend>control_msgs</exec_depend>
 
 </package>
