--- eager_bridge_gazebo/src/eager_bridge_gazebo/action_server/action_server.py
@@ -1,13 +1,0 @@
-import rospy
-import actionlib
-
-class ActionServer(object):
-    
-    def __init__(self, action_server, action_type):
-        self.client = actionlib.SimpleActionClient(action_server, action_type)
-        rospy.logdebug("Waiting for server {} ...".format(action_server))
-        self.client.wait_for_server()
-        rospy.logdebug("[{}] Connected to server".format(action_server))
-    
-    def act(self, action):
-        self.client.send_goal(action)
