+++ eager_bridge_real/package.xml
@@ -27,9 +27,5 @@
   <exec_depend>std_msgs</exec_depend>
   <exec_depend>roslaunch</exec_depend>
   <exec_depend>sensor_msgs</exec_depend>
-  <exec_depend>control_msgs</exec_depend>
-	<exec_depend>trajectory_msgs</exec_depend>
-	<exec_depend>actionlib</exec_depend>
-	<exec_depend>control_msgs</exec_depend>
 
 </package>
