+++ eager_bridge_real/src/eager_bridge_real/real_bridge.py
@@ -2,43 +2,17 @@ import rospy
 import roslaunch
 import functools
 import sensor_msgs.msg
-from operator import add
-from inspect import getmembers, isclass, isroutine
 import eager_core.action_server
+from inspect import getmembers, isclass, isroutine
 from eager_core.physics_bridge import PhysicsBridge
 from eager_core.utils.file_utils import substitute_xml_args
-from eager_bridge_gazebo.orientation_utils import quaternion_multiply, euler_from_quaternion
-from gazebo_msgs.srv import GetPhysicsProperties, GetPhysicsPropertiesRequest, SetPhysicsProperties, SetPhysicsPropertiesRequest
-from eager_bridge_gazebo.srv import SetInt, SetIntRequest
 from eager_core.utils.message_utils import get_value_from_def, get_message_from_def, get_response_from_def, get_length_from_def
 
-class GazeboBridge(PhysicsBridge):
+class RealBridge(PhysicsBridge):
 
     def __init__(self):
-        self._start_simulator()
-
-        step_time = rospy.get_param('physics_bridge/step_time', 0.1)
-
-        physics_parameters_service = rospy.ServiceProxy('/gazebo/get_physics_properties', GetPhysicsProperties)
-        physics_parameters_service.wait_for_service()
-
-        physics_parameters = physics_parameters_service(GetPhysicsPropertiesRequest())
-
-        new_physics_parameters = SetPhysicsPropertiesRequest()
-        new_physics_parameters.time_step = rospy.get_param('physics_bridge/solver_time_step', 0.001)
-        new_physics_parameters.max_update_rate = rospy.get_param('physics_bridge/solver_max_update_rate', 0.0)
-        new_physics_parameters.gravity = physics_parameters.gravity
-        new_physics_parameters.ode_config = physics_parameters.ode_config
-
-        set_physics_properties_service = rospy.ServiceProxy('/gazebo/set_physics_properties', SetPhysicsProperties)
-        set_physics_properties_service.wait_for_service()
-
-        set_physics_properties_service(new_physics_parameters)
-
-        self._step_world = rospy.ServiceProxy('/gazebo/step_world', SetInt)
-        self._step_world.wait_for_service()
-
-        self.step_request = SetIntRequest(int(round(step_time/physics_parameters.time_step)))
+        action_rate = rospy.get_param('physics_bridge/action_rate', 30)
+        self.rate = rospy.Rate(action_rate)
 
         self._sensor_buffer = dict()
         self._sensor_subscribers = []
@@ -50,22 +24,7 @@ class GazeboBridge(PhysicsBridge):
         self._state_subscribers = []
         self._state_services = []
 
-        super(GazeboBridge, self).__init__("gazebo")
-
-    def _start_simulator(self):
-        str_launch_sim = '$(find eager_bridge_gazebo)/launch/gazebo_sim.launch'
-        cli_args = [substitute_xml_args(str_launch_sim),
-                    'no_gui:=%s' % rospy.get_param('physics_bridge/no_gui', False),
-                    'world:=%s' % rospy.get_param('physics_bridge/world')]
-        seed = rospy.get_param('physics_bridge/seed', None)
-        if seed is not None:
-            cli_args.append('extra_gazebo_args:=--seed %d' % seed)
-        roslaunch_args = cli_args[1:]
-        roslaunch_file = [(roslaunch.rlutil.resolve_launch_arguments(cli_args)[0], roslaunch_args)]
-        uuid = roslaunch.rlutil.get_or_generate_uuid(None, False)
-        roslaunch.configure_logging(uuid)
-        self._launch = roslaunch.parent.ROSLaunchParent(uuid, roslaunch_file)
-        self._launch.start()
+        super(RealBridge, self).__init__("real")
 
     def _register_object(self, topic, name, package, object_type, args, config):
         self._add_robot(topic, name, package, args, config)
@@ -81,19 +40,11 @@ class GazeboBridge(PhysicsBridge):
         return True
 
     def _add_robot(self, topic, name, package, args, config):
-        if 'default_translation' in config:
-            pos = "-x {:.2f} -y {:.2f} -z {:.2f}".format(*map(add, config['default_translation'], args['position']))
-        else:
-            pos = "-x {:.2f} -y {:.2f} -z {:.2f}".format(*args['position'])
-        if 'default_orientation' in config:
-            ori = "-R {:.2f} -P {:.2f} -Y {:.2f}".format(*euler_from_quaternion(*quaternion_multiply(config['default_orientation'], args['orientation'])))
-        else:
-            ori = "-R {:.2f} -P {:.2f} -Y {:.2f}".format(*euler_from_quaternion(*args['orientation']))
-        str_launch_object = '$(find %s)/launch/gazebo.launch' % package
+        str_launch_object = '$(find %s)/launch/real.launch' % package
         cli_args = [substitute_xml_args(str_launch_object),
                     'ns:=%s' % topic,
                     'name:=%s' % name,
-                    'configuration:=%s' % (pos + " " + ori)]
+                    ]
         roslaunch_args = cli_args[1:]
         roslaunch_file = [(roslaunch.rlutil.resolve_launch_arguments(cli_args)[0], roslaunch_args)]
         uuid = roslaunch.rlutil.get_or_generate_uuid(None, False)
@@ -186,7 +137,7 @@ class GazeboBridge(PhysicsBridge):
                 (get_action_srv, set_action_srv) = self._actuator_services[robot][actuator]
                 actions = get_action_srv()
                 set_action_srv(actions.value)
-        self._step_world(self.step_request)
+        self.rate.sleep()
         return True
 
     def _reset(self):
